+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - update/delete activities
 - activity's notifications
 - remember sign in with email/password
+- closest events at activities page
 
 ## [1.1.9] - 2019-12-24
 ### Modified